+++ extensions/gif-search/src/components/GifActions.tsx
@@ -111,13 +111,13 @@ export function GifActions({ item, showViewDetails, visitGifItem, mutate }: GifA
 
   async function copyGifAsSquare() {
     try {
-      await showToast({ style: Toast.Style.Animated, title: "Copying GIF as Square" });
+      await showToast({ style: Toast.Style.Animated, title: "Copying GIF Square" });
       const file = await copyGifAsSquareToClipboard(item.gif_url, item.download_name);
       await trackUsage();
       await closeMainWindow();
       await showToast({ style: Toast.Style.Success, title: `Copied square GIF "${path.basename(file)}"` });
     } catch (error) {
-      await showFailureToast(error, { title: "Could not copy GIF as square" });
+      await showFailureToast(error, { title: "Could not copy GIF square" });
     }
   }
 
@@ -133,6 +133,20 @@ export function GifActions({ item, showViewDetails, visitGifItem, mutate }: GifA
     } catch (error) {
       console.error(error);
       await showFailureToast(error, { title: "Could not paste GIF" });
+    }
+  }
+
+  async function pasteGifSquare() {
+    try {
+      await showToast({ style: Toast.Style.Animated, title: "Pasting GIF Square" });
+      const file = await copyGifAsSquareToClipboard(item.gif_url, item.download_name);
+      await trackUsage();
+      await closeMainWindow();
+      await Clipboard.paste({ file });
+      await showToast({ style: Toast.Style.Success, title: `Pasted square GIF "${path.basename(file)}"` });
+    } catch (error) {
+      console.error(error);
+      await showFailureToast(error, { title: "Could not paste GIF square" });
     }
   }
 
@@ -224,6 +238,18 @@ export function GifActions({ item, showViewDetails, visitGifItem, mutate }: GifA
       onPaste={trackUsage}
     />
   );
+  const copySquareGif =
+    process.platform === "darwin" ? (
+      <Action icon={Icon.Clipboard} key="copySquareGif" title="Copy GIF Square" onAction={copyGifAsSquare} />
+    ) : (
+      <Action icon={Icon.Clipboard} key="copySquareGif" title="Copy GIF Square (macOS Only)" />
+    );
+  const pasteSquareGif =
+    process.platform === "darwin" ? (
+      <Action icon={Icon.Clipboard} key="pasteSquareGif" title="Paste GIF Square" onAction={pasteGifSquare} />
+    ) : (
+      <Action icon={Icon.Clipboard} key="pasteSquareGif" title="Paste GIF Square (macOS Only)" />
+    );
   const pasteGifUrl = (
     <Action.Paste
       key="pasteGifUrl"
@@ -287,21 +313,6 @@ export function GifActions({ item, showViewDetails, visitGifItem, mutate }: GifA
       onPush={trackUsage}
     />
   );
-  const copySquareGif =
-    process.platform === "darwin" ? (
-      <Action
-        icon={Icon.Crop}
-        key="copySquareGif"
-        title="Copy GIF as Square"
-        onAction={copyGifAsSquare}
-        shortcut={{
-          macOS: { modifiers: ["cmd", "shift"], key: "s" },
-          Windows: { modifiers: ["ctrl", "shift"], key: "s" },
-        }}
-      />
-    ) : (
-      <Action icon={Icon.Crop} key="copySquareGif" title="Copy GIF as Square (macOS Only)" />
-    );
 
   const copyPageUrl = url ? (
     <Action.CopyToClipboard
@@ -337,8 +348,8 @@ export function GifActions({ item, showViewDetails, visitGifItem, mutate }: GifA
   );
 
   const actions: Array<(React.JSX.Element | undefined)[]> = [
-    [copyFile, pasteFile, copyGifUrl, pasteGifUrl, copyGifMarkdown, pasteGifMarkdown],
-    [toggleFav, removeRecent, showViewDetails ? viewDetails : undefined, copySquareGif],
+    [copyFile, pasteFile, copyGifUrl, pasteGifUrl, copyGifMarkdown, pasteGifMarkdown, copySquareGif, pasteSquareGif],
+    [toggleFav, removeRecent, showViewDetails ? viewDetails : undefined],
     [copyPageUrl, openUrlInBrowser, downloadFileAction],
   ];